# Use accessor instead of copying memory in CPU nms_kernel