As discussed with @fmassa, here's a quick attempt to use a tensor accessor instead of pointers over continuous (copied) columns in the `nms` kernel.

According to a simple benchmark, this doesn't seem to lead to an improvement so I'll close the PR:

This PR:
```
row-aligned boxes
138 ms ± 2.61 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
col-aligned boxes
144 ms ± 2.32 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

Master:
```
row-aligned boxes
138 ms ± 1.07 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
col-aligned boxes
126 ms ± 903 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

<details>
<summary>Benchmark code</summary>

```py
import torch
from torchvision import ops
from IPython import get_ipython
ipython = get_ipython()

def create_tensors_with_iou(N, iou_thresh):
    boxes = torch.rand(N, 4) * 100
    boxes[:, 2:] += boxes[:, :2]
    boxes[-1, :] = boxes[0, :]
    x0, y0, x1, y1 = boxes[-1].tolist()
    iou_thresh += 1e-5
    boxes[-1, 2] += (x1 - x0) * (1 - iou_thresh) / iou_thresh
    scores = torch.rand(N)
    return boxes, scores

iou_threshold = .8
n_boxes = 10_000
boxes, scores = create_tensors_with_iou(n_boxes, iou_threshold)

print("row-aligned boxes")
ipython.magic("timeit ops.nms(boxes, scores, iou_threshold)")

boxes_F = torch.empty_strided((n_boxes, 4), stride=(1, n_boxes))
boxes_F[:] = boxes
print("col-aligned boxes")
ipython.magic("timeit ops.nms(boxes_F, scores, iou_threshold)")
```


</details>